Settings: Redesign Fluid Version
What we do here is introduce our new Layout as a LayoutPreference which is a part of AOSP Settings. Therefore we nuke the preferences related to the Fluid Version and Specs, use the FluidVersionPreferenceController for the new Layout and use the SpecsPreferenceController as a new Util which provides the static code for device-related info such as RAM, battery capacity, screen specifications etc.

Add this to your device tree:

PRODUCT_PRODUCT_PROPERTIES += \
  ro.fluid.maintainer=maintainername \
  ro.fluid.cpu=for example SDM855

Co-authored-by: Androbuddy <anirudhchimpidi8@gmail.com>